requirements.txt
python-dotenv
requests
pydantic


paddlepaddle-gpu
paddleocr
opencv-python
Pillow
numpy
scipy

pdf2image
pypdf
python-multipart

tqdm

fastapi
uvicorn

loguru
jsonschema
